MVP Windows Production Build (#7)
* Begin formatting with Material UI

I like it so far

* Continue Material UI Development

* Continue Material UI Development

* Major State Overhaul

Major refactor of backend state, utilizing an enum to ensure compiler-enforced code correctness.

Implemented save and load methods. The save method works, the load method is having permissions issues.

* Continue UI Development

* Refactor forms to not break

Simplify datetime handling of forms and correctly handle controlled vs. uncontrolled components with material UI.

* Update README.md

* Implemented Newtype Pattern

Used derive_more crate to add some impls I didn't want to manually do.

* Reformat Backend Code

Just add my comment dividers to make this thing easier to read because it was getting to be a bit much.

* Fix bugs in is modified state check

I forgot to have the backend emit that the state was changed... It's working now :)

* Add Toast Context Provider

Add a toast context provider to send short popup messages to the user.

* Refactor Global State

See the readme on frontend architecture for the design. Toasts were added as well to indicate when errors occurred.

* Update toast handling

Make toasts stack and add a toast queue.

* Halfway working, need to fix bugs

* Most bugs fixed

* Add dark mode

* Add Additional Plot Settings

* Pivot to apache echarts (#4)

* Pivoted to Apache ECharts

This was the best decision I could have ever made.

* Finalize Echarts Pivot

* Create production build (#6)

* Pivot to apache echarts (#4) (#5)

* Pivoted to Apache ECharts

This was the best decision I could have ever made.

* Finalize Echarts Pivot

* MVP Production Build Successful